requirements.txt
tensorflow-gpu
tensorflow
tqdm
pathlib
pandas
numpy
matplotlib
seaborn
pyyaml
opencv-python
Pillow
streamlit
jupyterlab
scikit-learn
scikit-image
sklearn
ffmpeg
pytest
pytest-dependency
pre-commit
dvc[gs]
apache-airflow>=2.2.3
mlflow>=1.22.0
pyngrok